backend/requirements.txt
# fastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AI
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
pillow==10.1.0

# Satellite imagery and geospatial
rasterio==1.3.8
geopandas==0.14.0
shapely==2.0.2
folium==0.15.0

# climate and weather APIs
requests==2.31.0
httpx==0.25.2

# PDF generation
reportlab==4.0.7
fpdf2==2.7.6

# data processing
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# database 
sqlalchemy==2.0.23
alembic==1.13.1

# configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2